Move snippet to vendor to prevent inline scripts
As @mrloop pointed out, the old method of loading the snippet
included injected an inline script tag into the document head. This
requires an `unsafe-inline` permission if CSP headers are present
which is a security bummer.

This approach (credit to @mrloop for it) injects the script into
the vendor.js file, avoiding the inline script tag and CSP issues.