Don't break Fastboot
1. Using the new Fastboot standards, don't run the snippet
if Fastboot isn't defined, because the snippet requires 'document'.
2. Reference Rollbar using window.Rollbar in the vendor shim,
because 'Rollbar' is undefined in Fastboot.
3. Add a very basic test to test these.

Note this doesn't do much besides ensure that the app doesn't break
when Fastboot is enabled. The vendor shim won't work, and since
window.Rollbar won't be defined, it won't actually forward any of the calls.